pyproject.toml
[project]
name = "linkmonk-rss"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "click>=8.1.8",
    "feedparser>=6.0.11",
    "httpx>=0.28.1",
    "jinja2>=3.1.5",
    "python-dotenv>=1.0.1",
]
